requirements.txt
streamlit
fastapi
uvicorn
langchain
langchain-groq
langchain-huggingface
langchain-community
langchain-text-splitters
chromadb
python-dotenv
pypdf
sentence-transformers
requests
